course_examples/requirements.txt
opencv-python
ultralytics
roboflow
python-dotenv
pyyaml

